文档网站加入首屏渲染功能

+++ build/doc.js
@@ -4,7 +4,8 @@ shelljs.env.NODE_ENV = 'production'
 const websiteProject = './zen0822.github.io'
 
 module.exports = function ({
-  appName = 'example'
+  appName = 'example',
+  push = false
 } = {}) {
   const path = require('path')
   const config = require('./config')
@@ -33,30 +34,32 @@ module.exports = function ({
       chunkModules: false
     }) + '\n')
 
-    shelljs.rm('-rf', websiteProject)
-    if (shelljs.exec('git clone https://github.com/zen0822/zen0822.github.io.git').code === 0) {
-      shelljs.echo('Git clone zen0822.github.io Success')
+    if (push) {
+      shelljs.rm('-rf', websiteProject)
+      if (shelljs.exec('git clone https://github.com/zen0822/zen0822.github.io.git').code === 0) {
+        shelljs.echo('Git clone zen0822.github.io Success')
 
-      shelljs.rm('-rf', `${websiteProject}/static`)
-      shelljs.cp('-r', `${config.doc.assetsRoot}/*`, `${websiteProject}`)
-      shelljs.echo(`${assetsPath} successfully copy to ${websiteProject}`)
+        shelljs.rm('-rf', `${websiteProject}/static`)
+        shelljs.cp('-r', `${config.doc.assetsRoot}/*`, `${websiteProject}`)
+        shelljs.echo(`${assetsPath} successfully copy to ${websiteProject}`)
 
-      // TODO: 准备解析 log 到网站分支
-      // let log = shelljs.exec('git log')
+        // TODO: 准备解析 log 到网站分支
+        // let log = shelljs.exec('git log')
 
-      shelljs.cd('./zen0822.github.io')
+        shelljs.cd('./zen0822.github.io')
 
-      shelljs.exec('git add -A')
-      shelljs.exec('git commit -m "更新网站"')
+        shelljs.exec('git add -A')
+        shelljs.exec('git commit -m "更新网站"')
 
-      shelljs.exec('git push origin master', function (code) {
-        code === 0 && console.log('Success: push to zen0822.github.io')
-        shelljs.cd('../')
-        shelljs.rm('-rf', './zen0822.github.io')
-      })
-    } else {
-      shelljs.echo('Error: Git clone zen0822.github.io failed')
-      exit(1)
+        shelljs.exec('git push origin master', function (code) {
+          code === 0 && console.log('Success: push to zen0822.github.io')
+          shelljs.cd('../')
+          shelljs.rm('-rf', './zen0822.github.io')
+        })
+      } else {
+        shelljs.echo('Error: Git clone zen0822.github.io failed')
+        exit(1)
+      }
     }
   })
 }
